finish off the tests' import issues and get a bit of separation; all tests passing on Python2.7
